pyproject.toml
[project]
name = "anydoor-refiners"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel Rodriguez", email = "darc2203@gmail.com" }
]
dependencies = [
    "refiners[conversion] @ git+https://github.com/finegrain-ai/refiners",
    "einops>=0.8.0",
    "ipython>=8.29.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "tqdm>=4.66.5",
    "pytest>=8.3.3",
    "omegaconf>=2.3.0",
    "pytorch-lightning>=2.4.0",
    "opencv-python>=4.10.0.84",
    "safetensors>=0.4.5",
    "matplotlib>=3.9.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/anydoor_refiners"]
